+++ packages/fusion/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./types.js";
+export * from "./recipe.js";
+export * from "./runner.js";
+export { registerFusion } from "./extension.js";